refactor(mediawiki): split links.go into check/internal siblings (deslop #59)
